The `process_map` function felt a bit slow, so I rewrote parts of it to try and speed it up. Based on the `profvis` profiling output, I replaced the `dplyr` style slicing with data.table methods, but kept the output exactly the same (column order, sorting).  

I used `microbenchmark` to runs `process_map(data)` 50 times and report how much less time it took (median time) to run the new version compared to the old. The resulting speed up is dependent on the data set:
* `sepsis` takes ~33% less time
* `patients` takes ~34% less time 
* `hospital_billings` takes ~70% less time (from 1525 ms to 459ms)
* `traffic_fines` takes ~72% less time (from 1457ms to 395ms)

On my real life dataset containing almost 200,000 rows it takes ~30% less time to run `process_map(data, frequency('absolute-case'))` (from 32s to 21s). The benchmark on the real life dataset was run on a different device than the example datasets.