@gertjanssenswillen 2025-07-07
Thanks for the updates. I'm doing some checks on dev, merge with master and publish on CRAN.  (also edeaR later). 
As a token of appreciation, I can add you as a contributor with your github username, or if your actual name if you let me know.